One Shared File Per Circuit to Optimize
