+++ AutoCkt/Auto/autockt/__init__.py
@@ -1,0 +1,5 @@
+"""
+# `autockt` Namespace
+"""
+
+from .train import train
