+++ AutoCkt/Shared/autockt_shared/opamp_output.py
@@ -1,0 +1,12 @@
+from dataclasses import dataclass
+
+
+@dataclass
+class OpAmpOutput:
+    """# Op-Amp Output
+    Server output type reused by several op-amp flavors"""
+
+    gain: float
+    ugbw: float
+    phm: float
+    ibias: float
