+++ AutoCkt/Shared/autockt_shared/rewards.py
@@ -1,17 +1,10 @@
-from autockt_shared import (
-    AutoCktOutput,
-    FlipFlopOutput,
-    FoldedCascodeOutput,
-    LDOOutput,
-    LatchOutput,
-    TwoStageOpAmpNgmOutput,
-    Number,
-)
+from .opamp_output import OpAmpOutput
+from .typing import Number
 
 
 # TODO: remove target_output, can be assumed from rl model
 def settaluri_reward(
-    curr_output: AutoCktOutput,  # This is from simulation
+    curr_output: OpAmpOutput,  # This is from simulation
     target_output: dict[str, Number],  # This is from Specs
 ):
     """
@@ -47,161 +40,6 @@ def settaluri_reward(
 
         # TODO this 10 seems pretty arbitrary
         return reward if reward < -0.02 else 10
-
-    # run the reward function
-    return reward(curr_output, target_output)
-
-
-def latch_reward(
-    curr_output: LatchOutput,  # This is from simulation
-    target_output: dict[str, Number],  # This is from Specs
-):
-    """
-    TODO Implement
-    """
-
-    def calc_relative(curr: Number, ideal: Number):
-        # ideal = float(ideal)  # Not sure if this is necessary
-        relative = curr / ideal
-        return relative
-
-    # adapted TwoAmp reward using new variables
-    def reward(curr_output, target_output):
-        # populate relative for each key input of target
-        pos_val = []
-        reward = 1.0
-        for key in target_output:
-            reward = reward * calc_relative(
-                getattr(curr_output, key),
-                target_output[key],
-            )
-
-        # TODO this 10 seems pretty arbitrary
-        return reward
-
-    # run the reward function
-    return reward(curr_output, target_output)
-
-
-def flip_flop_reward(
-    curr_output: FlipFlopOutput,  # This is from simulation
-    target_output: dict[str, Number],  # This is from Specs
-):
-    """
-    TODO Implement
-    """
-
-    def calc_relative(curr: Number, ideal: Number):
-        # ideal = float(ideal)  # Not sure if this is necessary
-        relative = curr / ideal
-        return relative
-
-    # adapted TwoAmp reward using new variables
-    def reward(curr_output, target_output):
-        # populate relative for each key input of target
-        pos_val = []
-        reward = 1.0
-        for key in target_output:
-            reward = reward * calc_relative(
-                getattr(curr_output, key),
-                target_output[key],
-            )
-
-        # TODO this 10 seems pretty arbitrary
-        return reward
-
-    # run the reward function
-    return reward(curr_output, target_output)
-
-
-def folded_cascode_reward(
-    curr_output: FoldedCascodeOutput,  # This is from simulation
-    target_output: dict[str, Number],  # This is from Specs
-):
-    """
-    TODO Implement
-    """
-
-    def calc_relative(curr: Number, ideal: Number):
-        # ideal = float(ideal)  # Not sure if this is necessary
-        relative = curr / ideal
-        return relative
-
-    # adapted TwoAmp reward using new variables
-    def reward(curr_output, target_output):
-        # populate relative for each key input of target
-        pos_val = []
-        reward = 1.0
-        for key in target_output:
-            reward = reward * calc_relative(
-                getattr(curr_output, key),
-                target_output[key],
-            )
-
-        # TODO this 10 seems pretty arbitrary
-        return reward
-
-    # run the reward function
-    return reward(curr_output, target_output)
-
-
-def ldo_reward(
-    curr_output: LDOOutput,  # This is from simulation
-    target_output: dict[str, Number],  # This is from Specs
-):
-    """
-    TODO Implement
-    """
-
-    def calc_relative(curr: Number, ideal: Number):
-        # ideal = float(ideal)  # Not sure if this is necessary
-        relative = curr / ideal
-        return relative
-
-    # adapted TwoAmp reward using new variables
-    def reward(curr_output, target_output):
-        # populate relative for each key input of target
-        pos_val = []
-        reward = 1.0
-        for key in target_output:
-            reward = reward * calc_relative(
-                getattr(curr_output, key),
-                target_output[key],
-            )
-
-        # TODO this 10 seems pretty arbitrary
-        return reward
-
-    # run the reward function
-    return reward(curr_output, target_output)
-
-
-def two_stage_op_amp_ngm_reward(
-    curr_output: TwoStageOpAmpNgmOutput,  # This is from simulation
-    target_output: dict[str, Number],  # This is from Specs
-):
-    """
-    TODO Implement
-    """
-
-    def calc_relative(curr: Number, ideal: Number):
-        # ideal = float(ideal)  # Not sure if this is necessary
-        relative = curr / ideal
-        return relative
-
-    # adapted TwoAmp reward using new variables
-    def reward(curr_output, target_output):
-        # populate relative for each key input of target
-        pos_val = []
-        reward = 1.0
-        for key in target_output:
-            reward = reward * calc_relative(
-                getattr(curr_output, key),
-                target_output[key],
-            )
-
-        # TODO this 10 seems pretty arbitrary
-        return reward
 
     # run the reward function
     return reward(curr_output, target_output)
