+++ AutoCkt/scripts/train_opamp.py
@@ -1,4 +1,4 @@
-from autockt_shared.latch import circuit_optimization
+from autockt_shared.opamp import circuit_optimization
 from autockt import train
 
 train(circuit_optimization)
