+++ scripts/speed_multiple.py
@@ -9,10 +9,8 @@ if __name__ != "__main__":
 
 
 # Local Imports
-from autockt.trainer import (
-    autockt_train,
-)
-from autockt.autockt_gym_env_config import (
+from autockt.trainer import autockt_train
+from autockt_shared.cktopt import (
     AutoCktCircuitOptimization,
     AutoCktGymEnvConfig,
     AutoCktParams,
@@ -21,8 +19,8 @@ from autockt.autockt_gym_env_config import (
     AutoCktSpec,
 )
 from autockt_shared import (
-    AutoCktInput,
-    AutoCktOutput,
+    OpAmpInput,
+    OpAmpOutput,
     auto_ckt_sim,
 )
 from autockt_shared.rewards import (
@@ -54,8 +52,8 @@ def main():
                 AutoCktSpec("ibias", (0.0001, 0.01), normalize=0.001),
             ]
         ),
-        input_type=AutoCktInput,
-        output_type=AutoCktOutput,
+        input_type=OpAmpInput,
+        output_type=OpAmpOutput,
         simulation=auto_ckt_sim,
         reward=settaluri_reward,
     )
