+++ tests/test_auto_ckt_sim_hdl21.py
@@ -1,9 +1,8 @@
 from autockt_shared import (
     auto_ckt_sim_hdl21,
-    AutoCktInput,
-    AutoCktOutput,
+    OpAmpInput,
 )
 
-inp = AutoCktInput(3, 3, 3, 3, 3, 3, 1e-12)
+inp = OpAmpInput(3, 3, 3, 3, 3, 3, 1e-12)
 out = auto_ckt_sim_hdl21(inp)
 print(out)
